权限中心 字典管理

+++ toucan-shopping-modules/adminAuth/toucan-shopping-admin-auth-entity-vo/src/main/java/com/toucan/shopping/modules/admin/auth/entity/Dict.java
@@ -28,7 +28,7 @@ public class Dict {
 
     private String code; //编码
 
-    private Integer pid; //父级ID
+    private Long pid; //父级ID
 
     private Integer categoryId; //字典分类ID
